the `--max-targets` flag with a percent wasn't respecting multi-ports (#886)
* added example of max targets percentage

* max hosts percent for multi port, push to test

* WIP

* added logging

* added logging p2

* fixed bug with type cast

* made .h and .c consistent with port_count naming

* removed extraneous log

* edited --help

* fixed compile warning

* use uint_64 to be explicit

* missed a uint64_t

* added warn message if max-targets greater than search space

* removed log

* added integration test for max-targets

+++ src/state.h
@@ -166,7 +166,7 @@ struct state_send {
 	int warmup;
 	int complete;
 	uint32_t first_scanned;
-	uint32_t max_targets;
+	uint64_t max_targets;
 	uint32_t sendto_failures;
 	uint32_t max_index;
 	uint16_t max_port_index;
@@ -215,7 +215,7 @@ struct state_recv {
 extern struct state_recv zrecv;
 
 struct port_conf {
-	int port_count;
+	uint port_count;
 	uint16_t ports[0xFFFF + 1];
 	uint8_t *port_bitmap;
 };
